Minimum length correction Add enrollment validation to the op_student Minimum length correction

+++ myschool/op_student/op_student.py
@@ -675,17 +675,18 @@ class op_student(osv.Model):
                 c = b[2]
                 if c is not False:
                     d = c.get('batch_code')
-                    e.append(d)
-                    f = e.count(d)
-                    if f > 1:
-                        raise osv.except_osv (_('Course Enrollment Error'), _('Enrollments cannot be duplicated'))
-                    elif f == 1:
-                        stu = self.browse(cr, uid, ids, context).ids
-                        se = self.pool.get('op.enrollment')
-                        sa = se.search(cr, uid, ['&', ('batch_code', '=', d), ('student_id', '=', stu)], context=context)
-                        if len(sa) == 1:
-                            # raise osv.except_osv(_('Error'), _('All ready Exist'))
-                            cr.execute('Delete from op_enrollment where id =%s' % sa[0])
+                    if d:
+                        e.append(d)
+                        f = e.count(d)
+                        if f > 1:
+                            raise osv.except_osv (_('Course Enrollment Error'), _('Enrollments cannot be duplicated'))
+                        elif f == 1:
+                            stu = self.browse(cr, uid, ids, context).ids
+                            se = self.pool.get('op.enrollment')
+                            sa = se.search(cr, uid, ['&', ('batch_code', '=', d), ('student_id', '=', stu)], context=context)
+                            if len(sa) == 1:
+                                # raise osv.except_osv(_('Error'), _('All ready Exist'))
+                                cr.execute('Delete from op_enrollment where id =%s' % sa[0])
 
         return super(op_student, self).write(cr, uid, ids, values, context=context)
